+++ RELEASE_NOTES.md
@@ -4,6 +4,15 @@ Release Notes
 ## vNext
 * WebApps/Functions: Specify connection string types
 
+## 1.6.30
+* WebApps/Functions: Allow adding IP restriction string with CIDR
+
+## 1.6.29
+* CLI: include `--only-show-error` option when executing Azure CLI commands.
+
+## 1.6.28
+* ServicePlan/WebApp: Support for enabling ZoneRedundant
+
 ## 1.6.27
 * Functions: Make `connection_string` available for Azure Functions in addition to WebApps.
 * WebApps/Functions: Add support for ip-restriction rules